test: make relay concurrency regression detect miscounts without -race
Review follow-up. The regression indexed each relay under its full key in
byPathHop as well, which eviction does not compact, so a stale byNode
candidate never changed the bulk result and only the race detector saw the
bug. Relay evidence is now reachable only via byNode (restart shape): on
fa50931d the oracle check fails without -race too. Also fixes the eviction
cutoff comment (keepTx+1 remain) and uses int for candidate indexes.

+++ cmd/server/repeater_enrich_bulk.go
@@ -144,7 +144,7 @@ func (s *PacketStore) computeRepeaterRelayInfoMap(windowHours float64) map[strin
 	// was indexed above under the same lock); it is skipped, never mapped to
 	// another transmission.
 	type relayRef struct {
-		tx         int32
+		tx         int
 		fromPrefix bool
 	}
 	type keyRefs struct {
@@ -171,7 +171,7 @@ func (s *PacketStore) computeRepeaterRelayInfoMap(windowHours float64) map[strin
 				b.gen = gen
 				for _, have := range b.keys {
 					if have == key {
-						refs = append(refs, relayRef{tx: int32(i), fromPrefix: fromPrefix})
+						refs = append(refs, relayRef{tx: i, fromPrefix: fromPrefix})
 						return
 					}
 				}
